aula 15-03-24

+++ database/migrations/2024_03_01_222719_create_reservas_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->integer('idfuncionario');
             $table->integer('idquarto');
             $table->enum('situacao', ['Pago','Pendente']);
-            $table->double('valortotal', 8,2);
+            $table->decimal('valortotal', 8,2);
             $table->date('dataentrada');
             $table->date('datasaida');
         });
